+++ app/views/layouts/application.html.haml
@@ -1,13 +1,13 @@
 !!!
 %html
   %head
-    %title Rotten Potatoes!
+    %title Website Title!
     = stylesheet_link_tag 'application', 'media' => 'all'
     = javascript_include_tag 'application'
     = csrf_meta_tags
 
   %body
-    %h1.title Rotten Potatoes!
+    %h1.title Website Title!
     #main
       - if flash[:notice]
         #notice.message= flash[:notice]
