+++ app/views/sessions/new.html.haml
@@ -1,0 +1,23 @@
+%h1 Login
+
+= form_tag movies_path do
+
+  = label :movie, :title, 'Title'
+  = text_field :movie, 'title'
+
+  = label :movie, :rating, 'Rating'
+  = select :movie, :rating, ['G','PG','PG-13','R','NC-17']
+
+  = label :movie, :release_date, 'Released On'
+  = date_select :movie, :release_date
+
+  = submit_tag 'Save Changes'
+
+= form_with(scope: :user, url: session_path(:user), local: true) do |form|
+  = form.label :email
+  = form.email_field :email, required: true
+
+  = form.label :password
+  = form.password_field :password, required: true
+
+  = form.submit "Log in"
