+++ spec/models/item_spec.rb
